// projects/xpx-typescript-angular-sdk/src/model/download-request.ts
import { MessageType } from './message-type';

/**
 * Copyright 2018 ProximaX Limited
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Interface represents the DownloadRequest
 */
export interface DownloadRequest {

    /**
     * The nem datahash
     */
    hash?: string;


    /**
     * The recipient private key
     */
    recipientPrivateKey?: string;

    /**
     * The sender public key
     */
    senderPublicKey?: string;

    /**
     * The message type
     */
    messageType?: MessageType;

}